requirements.txt
flask==3.0.0
gunicorn==21.2.0
tensorflow==2.15.0
numpy==1.24.3
pillow==10.1.0
h5py==3.10.0
